# CORE-2005: folder listing performance improvements.