--- a/src/clj_icat_direct/queries.clj
+++ b/src/clj_icat_direct/queries.clj
@@ -180,6 +180,12 @@
                                       WHERE d2.data_id = d1.data_id)"))
 
 
+(defn mk-colls-in-coll
+  [parent-path]
+  (str "SELECT *
+          FROM r_coll_main AS c1
+          WHERE c1.parent_coll_name = '" parent-path "'"))
+
 (defn- mk-obj-avus
   [obj-ids-query]
   (str "SELECT object_id, meta_attr_value, meta_attr_name
@@ -236,7 +242,7 @@
 (defn- mk-folders-in-folder
   [parent-path group-ids-query include-count?]
   (str "SELECT 'collection'                           AS type,
-               m.meta_attr_value                      AS uuid,
+               ca.meta_attr_value                     AS uuid,
                c.coll_name                            AS full_path,
                REGEXP_REPLACE(c.coll_name, '.*/', '') AS base_name,
                NULL                                   AS info_type,
@@ -248,12 +254,11 @@
        (if include-count? ",
                COUNT(*) OVER () AS total_count" "")
          " FROM r_coll_main AS c
-            JOIN r_objt_metamap AS om ON om.object_id = c.coll_id
-            JOIN r_meta_main AS m ON m.meta_id = om.meta_id
+            JOIN coll_avus AS ca ON ca.object_id = c.coll_id
             JOIN r_objt_access AS a ON c.coll_id = a.object_id
           WHERE c.parent_coll_name = '" parent-path "'
             AND c.coll_type != 'linkPoint'
-            AND m.meta_attr_name = 'ipc_UUID'
+            AND ca.meta_attr_name = 'ipc_UUID'
             AND a.user_id IN (" group-ids-query ")
           GROUP BY type, uuid, full_path, base_name, info_type, data_size, c.create_ts,
                    c.modify_ts, data_checksum"))
@@ -581,12 +586,15 @@
       access_type_id - the ICAT DB Id indicating the user's level of access to the folder
       data_checksum  - nil for collections, the MD5 checksum of the file for data objects"
   [& {:keys [user zone parent-path sort-column sort-direction limit offset groups-table-query]}]
-  [[(str "WITH groups AS (" (or groups-table-query (mk-groups user zone)) ")
+  [[(mk-temp-table "colls" (mk-colls-in-coll parent-path))]
+   [(analyze "colls")]
+   [(mk-temp-table "coll_avus" (mk-obj-avus "SELECT coll_id FROM colls"))]
+   [(analyze "coll_avus")]
+   [(str "WITH groups AS (" (or groups-table-query (mk-groups user zone)) ")
        " (mk-folders-in-folder parent-path "SELECT group_user_id FROM groups" true) "
         ORDER BY " sort-column " " sort-direction "
         LIMIT ?
         OFFSET ?") limit offset]])
-
 
 (defn ^ISeq mk-paged-folder
   "This function constructs a set of parameterized queries for returning a sorted page of files and folders
@@ -629,6 +637,10 @@
      [(analyze "objs")]
      [(mk-temp-table "file_avus" (mk-obj-avus "SELECT data_id FROM objs"))]
      [(analyze "file_avus")]
+     [(mk-temp-table "colls" (mk-colls-in-coll parent-path))]
+     [(analyze "colls")]
+     [(mk-temp-table "coll_avus" (mk-obj-avus "SELECT coll_id FROM colls"))]
+     [(analyze "coll_avus")]
      [(str "WITH groups AS (" (or groups-table-query (mk-groups user zone)) ") "
            "SELECT *, COUNT(*) OVER () AS total_count
             FROM (" folders-query " UNION " files-query ") AS t