The purpose of this PR is to improve folder listings in some cases where the listing takes much longer than we would expect. The results were a little mixed, but the one case that we were trying to fix improved dramatically.

Before improvements:

| Case Description | Time |
| ---- | ---- |
| File and folder listing for problem account | 100 seconds |
| File and folder listing for non-problem account | 3.5 seconds |
| Folder listing only for problem account | 1.7 seconds |
| Folder listing only for non-problem account | 1.4 seconds |

After improvements:

| Case Description | Time |
| ---- | ---- |
| File and folder listing for problem account | 5.1 seconds |
| File and folder listing for non-problem account | 4.1 seconds |
| Folder listing only for problem account | 4.9 seconds |
| Folder listing only for non-problem account | 4.0 seconds |

The small performance decreases for everything but the problem case is suboptimal, but fixing the problem case is a higher priority at this time. We can spend some more time optimizing this some more later.