@ianmcorvidae 2024-07-16
One other thought and this is because I'm paranoid about these queries: did you try these out at all on more normal folders? (i.e. not the edge-case problem child that started this journey). Sometimes things can have weird consequences further away, though based on my gut feelings about this I doubt it would here; still maybe worth a quick check at least.