@slr71 2024-07-17
Doing this for some more normal folders is a good idea. Here are the times for both the problem case and a much more normal case: